합치는 과정 리뷰, 커뮤니티 관련 유저아이디 추가

+++ app/src/Cuminte/Cuminty.ctrl.js
@@ -52,16 +52,26 @@ const post = async (req, res) => {
 
 // 생성
 const createpost = async (req, res) => {
+  if (!req.session || !req.session.userId) {
+    return res.status(401).json({
+      resultCode: "F-2",
+      msg: "로그인이 필요합니다.",
+    });
+  }
+
   try {
-    const { post_title, post_content, post_date } = req.body;
-    console.log("Received Data:", req.body); // 입력 값 확인을 위한 로그
+    const { post_title, post_content } = req.body;
+    const username = req.session.username;
+    const userId = req.session.userId; // 세션에서 userId 가져오기
+
+    const post_date = new Date().toISOString().slice(0, 10);
 
     const query = `
-      INSERT INTO posts (title, content, post_date)
-      VALUES ($1, $2, $3)
+      INSERT INTO posts (title, content, post_date, username, author_id) 
+      VALUES ($1, $2, $3, $4, $5) 
       RETURNING *
     `;
-    const values = [post_title, post_content, post_date];
+    const values = [post_title, post_content, post_date, username, userId];
 
     const { rows } = await pool.query(query, values);
     console.log("Query Result:", rows); // 쿼리 결과 확인을 위한 로그
@@ -88,6 +98,14 @@ const createpost = async (req, res) => {
 };
 
 const remotepost = async (req, res) => {
+  if (!req.session || !req.session.userId) {
+    return res.status(401).json({
+      resultCode: "F-2",
+      msg: "로그인이 필요합니다.",
+    });
+  }
+  const userId = req.session.userId; // 로그인된 사용자 ID
+
   try {
     const { postId, post_title, post_content, post_date } = req.body;
 
@@ -97,11 +115,11 @@ const remotepost = async (req, res) => {
     const query = `
       UPDATE posts
       SET title = $1, content = $2, post_date = $3
-      WHERE post_id = $4
+      WHERE post_id = $4 AND userId = $5
       RETURNING *
     `;
 
-    const values = [post_title, post_content, actualPostDate, postId];
+    const values = [post_title, post_content, actualPostDate, postId, userId];
 
     const { rows } = await pool.query(query, values);
     console.log("Update Query Result:", rows);
@@ -129,11 +147,20 @@ const remotepost = async (req, res) => {
 
 // 삭제
 const deletepost = async (req, res) => {
+  if (!req.session || !req.session.userId) {
+    return res.status(401).json({
+      resultCode: "F-2",
+      msg: "로그인이 필요합니다.",
+    });
+  }
+
   try {
     const { post_id } = req.params;
+    const userId = req.session.userId; // 로그인된 사용자 ID
+
     const { rows } = await pool.query(
-      "DELETE FROM posts WHERE post_id = $1 RETURNING *",
-      [post_id]
+      "DELETE FROM posts WHERE post_id = $1 AND author_id = $2 RETURNING *",
+      [post_id, userId]
     );
     console.log("Delete Query Result:", rows); // 삭제 결과 확인을 위한 로그
